feat: print home directory after each operation
