+++ src/zlib.js
@@ -4,18 +4,14 @@ import { pipeline } from 'stream/promises';
 
 class Zlib {
   async compress(command, path, newPath) {
-    try {
-      const fileReadStream = createReadStream(path);
-      const fileWriteStream = createWriteStream(newPath);
-      const brotli =
-        command === 'compress'
-          ? createBrotliCompress()
-          : createBrotliDecompress();
-      await pipeline(fileReadStream, brotli, fileWriteStream);
-      console.log(`File ${command}ed`);
-    } catch (error) {
-      console.error('Operation failed', error.message);
-    }
+    const fileReadStream = createReadStream(path);
+    const fileWriteStream = createWriteStream(newPath);
+    const brotli =
+      command === 'compress'
+        ? createBrotliCompress()
+        : createBrotliDecompress();
+    await pipeline(fileReadStream, brotli, fileWriteStream);
+    console.log(`File ${command}ed`);
   }
 }
